Refactoring: HospitalizationRepository moved to repository package
