+++ src/main/java/pl/sdaproject/medicalarchivemanagementsystem/repository/HospitalizationRepository.java
@@ -1,4 +1,4 @@
-package pl.sdaproject.medicalarchivemanagementsystem.dao;
+package pl.sdaproject.medicalarchivemanagementsystem.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
